project2/requirements.txt
Pyro5==5.14
pycryptodome==3.17
dataclasses-json==0.5.7
pytz==2023.3